Change font from Mulish to Atkinson Hyperlegible for main and Fraunces for accents. Change all font vars, mixins, stylings accordingly. Created folders for individual pages and components, moved page and component files to matching folders. Deleted component and page sections from /styles.
