+++ client/src/components/index.js
@@ -1,0 +1,5 @@
+export { default as DefaultLayout } from './DefaultLayout/DefaultLayout.jsx';
+export { default as Container } from './Container/Container.jsx';
+
+export { default as AppHeader } from './AppHeader/AppHeader.jsx';
+export { default as AppFooter } from './AppFooter/AppFooter.jsx';
