+++ client/src/pages/index.js
@@ -1,5 +1,5 @@
-export { default as HomePage } from './HomePage.jsx';
-export { default as PromptStep } from './PromptStep.jsx';
-export { default as NameStep } from './NameStep.jsx';
-export { default as AestheticStep } from './AestheticStep.jsx';
-export { default as SummaryPage } from './SummaryPage.jsx';
+export { default as HomePage } from './HomePage/HomePage.jsx';
+export { default as PromptStep } from './PromptStep/PromptStep.jsx';
+export { default as NameStep } from './NameStep/NameStep.jsx';
+export { default as AestheticStep } from './AestheticStep/AestheticStep.jsx';
+export { default as SummaryPage } from './SummaryPage/SummaryPage.jsx';
